+++ src/main/java/com/kookdonge/kookdonge_server/auth/interceptor/AuthInterceptor.java
@@ -4,8 +4,8 @@ import com.kookdonge.kookdonge_server.auth.common.AuthExceptionCode;
 import com.kookdonge.kookdonge_server.auth.common.JWT;
 import com.kookdonge.kookdonge_server.auth.service.UserService;
 import com.kookdonge.kookdonge_server.auth.service.annotation.LoginRequired;
-import com.kookdonge.kookdonge_server.common.CustomException;
-import com.kookdonge.kookdonge_server.common.UserInfoStore;
+import com.kookdonge.kookdonge_server.common.exception.CustomException;
+import com.kookdonge.kookdonge_server.common.info.UserInfoStore;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import java.util.Objects;
