+++ src/main/java/com/kookdonge/kookdonge_server/auth/presentation/UserPresentation.java
@@ -6,8 +6,8 @@ import com.kookdonge.kookdonge_server.auth.presentation.dto.res.RegisterUserRes;
 import com.kookdonge.kookdonge_server.auth.presentation.dto.res.ReissueAccessTokenRes;
 import com.kookdonge.kookdonge_server.auth.service.UserService;
 import com.kookdonge.kookdonge_server.auth.service.dto.RegisterUserDTO;
-import com.kookdonge.kookdonge_server.common.RequestDTO;
-import com.kookdonge.kookdonge_server.common.ResponseDTO;
+import com.kookdonge.kookdonge_server.common.dto.RequestDTO;
+import com.kookdonge.kookdonge_server.common.dto.ResponseDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
