+++ src/main/java/com/kookdonge/kookdonge_server/common/config/AwsS3Config.java
@@ -6,7 +6,7 @@ import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import com.kookdonge.kookdonge_server.common.AwsS3Property;
+import com.kookdonge.kookdonge_server.common.property.AwsS3Property;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
