+++ src/main/java/com/kookdonge/kookdonge_server/common/dto/RequestDTO.java
@@ -1,4 +1,4 @@
-package com.kookdonge.kookdonge_server.common;
+package com.kookdonge.kookdonge_server.common.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
