+++ src/main/java/com/kookdonge/kookdonge_server/common/exception/ExceptionCode.java
@@ -1,4 +1,4 @@
-package com.kookdonge.kookdonge_server.common;
+package com.kookdonge.kookdonge_server.common.exception;
 
 public interface ExceptionCode {
     String getMessage();
